chore: simplify code
One code to rule them all.

This code is pretty similar to the one @ldemailly provided for
https://github.com/fortio/safecast

We worked at the same time on the same code, we came to distinct solutions.
His one was better and simpler than mine, so now I'm simply switching to
something highly inspired from his code.

Co-Authored-By: Laurent Demailly <ldemailly@gmail.com>