chore: change the signature of the Convert and MustConvert
These functions were accepting `any` as argument, which was not really helpful.

Some types were not supported and the users had to figure out by himself at the runtime.
The unsupported types are now reported at the compilation time.

This is a breaking change, but it's the whole point of this commit.

BREAKING CHANGE: a minor one but still breaking one.